+++ lib_acl/src/code/acl_htmlcode.c
@@ -123,7 +123,7 @@ int acl_html_decode(const char *in, ACL_VSTRING *out)
 				&ptr, NULL, NULL);
 		if (token == NULL) {
 			pre = markup_unescape(pre, out);
-			len = ptr - pre;
+			len = (int) (ptr - pre);
 			if (len > 0)
 				acl_vstring_memcat(out, pre, len);
 			break;
@@ -131,7 +131,7 @@ int acl_html_decode(const char *in, ACL_VSTRING *out)
 		spec = (const HTML_SPEC*) token->ctx;
 		acl_assert(spec != NULL);
 
-		len = ptr - pre - spec->len;
+		len = (int) (ptr - pre - spec->len);
 		if (len > 0)
 			acl_vstring_memcat(out, pre, len);
 		if (spec->ch > 255)
